# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de recolección de residuos",
        "title": "API Recolecta",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/notificaciones-push/ciudadanos/difusion": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía una notificación push con título y cuerpo a todos los ciudadanos registrados en el sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PushNotification"
                ],
                "summary": "Enviar notificación general a todos los ciudadanos (Broadcast)",
                "parameters": [
                    {
                        "description": "Datos del mensaje general",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.broadcastMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de envío para cada ciudadano",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/domain.SendResult"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/notificaciones-push/ciudadanos/difusion/punto/{point_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía una notificación push con título y cuerpo a todos los ciudadanos con domicilios en un radio de 200m del punto de parada especificado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PushNotification"
                ],
                "summary": "Enviar notificación general a ciudadanos cerca de un punto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del Punto de Parada (ej. 15)",
                        "name": "point_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del mensaje general",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.broadcastMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de envío para cada ciudadano",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/domain.SendResult"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/notificaciones-push/ciudadanos/difusion/ruta/{route_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Envía una notificación push con título y cuerpo a todos los ciudadanos que tengan un domicilio registrado en la colonia de la ruta especificada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PushNotification"
                ],
                "summary": "Enviar notificación general a ciudadanos de una ruta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la Ruta",
                        "name": "route_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del mensaje general",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.broadcastMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de envío para cada ciudadano",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/domain.SendResult"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/puntos-recoleccion/ruta/{rutaId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PuntoRecoleccion"
                ],
                "summary": "Puntos de recolección por ruta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ruta",
                        "name": "rutaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.PuntoRecoleccionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/puntos-recoleccion/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PuntoRecoleccion"
                ],
                "summary": "Punto de recolección por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.PuntoRecoleccionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PuntoRecoleccion"
                ],
                "summary": "Actualizar punto de recolección",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.UpdatePuntoRecoleccionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.PuntoRecoleccionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PuntoRecoleccion"
                ],
                "summary": "Eliminar punto de recolección",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.PuntoRecoleccionMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "core.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/core.ErrorDetail"
                }
            }
        },
...
        "domain.SendResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.PuntoRecoleccion": {
            "type": "object",
            "properties": {
                "cp": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "eliminado": {
                    "type": "boolean"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "punto_id": {
                    "type": "integer"
                },
                "ruta_id": {
                    "type": "integer"
                }
            }
        },
...
        "entities.PuntoRecoleccionMessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "entities.PuntoRecoleccionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/entities.PuntoRecoleccion"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.UpdatePuntoRecoleccionRequest": {
            "type": "object",
            "properties": {
                "cp": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "punto_id": {
                    "type": "integer"
                },
                "ruta_id": {
                    "type": "integer"
                }
            }
        },
...
        "infrastructure.broadcastMessageRequest": {
            "type": "object",
            "required": [
                "body",
                "title"
            ],
            "properties": {
                "body": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe 'Bearer ' seguido de tu token JWT. Ejemplo: \"Bearer eyJhbG...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}